Fix more NDCubeSequence bugs.
When introducing the single dispatch for NDCubeSequence slicing, the cube_item input was removed from the functions addressing the cases where the input was an int or slice.  This led to bugs when the item was a tuple as the int and slice cases were called requiring the ability to define the cube_item input.  This commit fixes that problem.