Add tests and fix bug in cropping by quantity extra coord.
This PR adds tests for NDCube.crop_by_extra_coord and in doing so, fixes a bug which wrongly converted an extra coord value from a Quantity to an array.  Extra coords should only be converted to an array if they are not a Quantity so that array operations can be performed.  These operations can already be performed on quantities and converting them to an array means we lose the unit information.