Change API for some util sequence functions.
Made some util functions only used in slicing private.  These functions should not have to be testing explicitly as they should be tested in the process of testing the NDCubeSequence slicing.

Also, changed the input arguments of convert_cube_like_item_to_sequence_items() and its sub-routines so it does not require an NDCubeSequence, but rather just the infomation from the sequences it uses.