First draft of adopting astropy's WCS APE.
This commit does not the change all the uses of the now altered methods in tests.  Furthermore, the new world_axis_physical_types() method does not currently return strings consistent with the IOVA UCD1+ vocabulary, but rather the WCS axis types.

Finally, the NDCubeSequence.dimensions property now represents the length of all cubes along the common_axis as a tuple within the dimensions tuple, unless all cubes are the same length in which case as an int.